requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.46.1
starlette==0.38.5
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
python-multipart==0.0.9